requirements.txt
Flask
flasgger
mysql-connector-python
python-dotenv
huggingface-hub
gunicorn
groq
requests
python-dotenv